updates demos to work w/ other browsers
