+++ demos/0-background-clearing/index.html
@@ -21,9 +21,6 @@
 <body>
   <canvas></canvas>
 
-  <script src="../../assets/webgl-debug.js"></script>
-  <script src="../../assets/gl-matrix.min.js"></script>
-  <script src="../../assets/webgl-utils.js"></script>
-  <script src="index.js"></script>
+  <script src="../../dist/0-background-clearing.js"></script>
 </body>
 </html>
