+++ demos/1-rotating-squares/index.html
@@ -35,9 +35,7 @@
     gl_FragColor = vec4(1.0, 1.0, 1.0, 1.0);
   }
   </script>
-  <script src="../../assets/webgl-debug.js"></script>
-  <script src="../../assets/gl-matrix.min.js"></script>
-  <script src="../../assets/webgl-utils.js"></script>
-  <script src="squares.js"></script>
+
+  <script src="../../dist/1-rotating-squares.js"></script>
 </body>
 </html>
