+++ demos/2-circle-approximations/index.html
@@ -26,10 +26,6 @@ void main ()
 }
 </script>
 
-
-  <script src="../../assets/webgl-debug.js"></script>
-  <script src="../../assets/gl-matrix.min.js"></script>
-  <script src="../../assets/webgl-utils.js"></script>
-  <script src="index.js"></script>
+<script src="../../dist/2-circle-approximations.js"></script>
 </body>
 </html>
