+++ demos/3-triangles/index.html
@@ -46,9 +46,6 @@ void main ()
 
 </script>
 
-  <script src="../../assets/webgl-debug.js"></script>
-  <script src="../../assets/gl-matrix.min.js"></script>
-  <script src="../../assets/webgl-utils.js"></script>
-  <script src="index.js"></script>
+<script src="../../dist/3-triangles.js"></script>
 </body>
 </html>
