+++ demos/4-obj-reader/index.html
@@ -143,6 +143,6 @@ void main ()
 }
 </script>
 
-  <script src="../../dist/4-obj-reader.entry.js"></script>
+<script src="../../dist/4-obj-reader.js"></script>
 </body>
 </html>
