+++ demos/5-textures/index.html
@@ -45,9 +45,6 @@ void main ()
 
 </script>
 
-  <script src="../../assets/webgl-debug.js"></script>
-  <script src="../../assets/gl-matrix.min.js"></script>
-  <script src="../../assets/webgl-utils.js"></script>
-  <script src="index.js"></script>
+<script src="../../dist/5-textures.js"></script>
 </body>
 </html>
