+++ demos/6-cubes/index.html
@@ -70,6 +70,6 @@ void main ()
 }
 </script>
 
-<script src="../../dist/6-cubes.entry.js"></script>
+<script src="../../dist/6-cubes.js"></script>
 </body>
 </html>
