+++ webpack.config.js
@@ -2,12 +2,17 @@
 
 module.exports = {
   entry: {
-    '6-cubes': "./demos/6-cubes",
+    '0-background-clearing': "./demos/0-background-clearing",
+    '1-rotating-squares': "./demos/1-rotating-squares",
+    '2-circle-approximations': "./demos/2-circle-approximations",
+    '3-triangles': "./demos/3-triangles",
     '4-obj-reader': "./demos/4-obj-reader",
+    '5-textures': "./demos/5-textures",
+    '6-cubes': "./demos/6-cubes",
   },
   output: {
     path: __dirname + '/dist',
-    filename: '[name].entry.js',
+    filename: '[name].js',
   },
   module: {
     loaders: [
